requirements.txt
browser-use==0.1.0
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2

playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

asyncio
pydantic==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1

regex==2023.10.3

colorlog==6.8.0
python-json-logger==2.0.7
